Moved ngram order check to counter module
