+++ nltk/model/counter.py
@@ -12,7 +12,19 @@ from copy import copy
 from nltk.probability import FreqDist, ConditionalFreqDist
 from nltk import compat
 
-from nltk.model.util import check_ngram_order, default_ngrams
+from nltk.model.util import default_ngrams, POS_INF
+
+
+def check_ngram_order(order_to_check, max_order=POS_INF):
+    """Sanity-check ngram order number."""
+    if order_to_check < 1:
+        raise ValueError("Ngram order cannot be less than 1. " "Got: {0}".format(order_to_check))
+    if order_to_check == POS_INF:
+        raise ValueError("Ngram order cannot be infinity!")
+    if order_to_check > max_order:
+        raise ValueError("Ngram order cannot be greater than {0}. "
+                         "Got: {1}".format(max_order, order_to_check))
+    return order_to_check
 
 
 def count_ngrams(order, vocabulary, *training_texts):
