+++ nltk/test/unit/model/test_counter.py
@@ -10,7 +10,8 @@ import unittest
 from nltk import six
 
 from nltk.model import NgramModelVocabulary, NgramCounter
-from nltk.model.util import check_ngram_order, POS_INF, default_ngrams
+from nltk.model.counter import check_ngram_order
+from nltk.model.util import default_ngrams
 
 
 class NgramCounterTests(unittest.TestCase):
@@ -124,7 +125,7 @@ class CheckNgramOrderTests(unittest.TestCase):
 
     def test_pos_inf_input(self):
         with self.assertRaises(ValueError):
-            check_ngram_order(POS_INF)
+            check_ngram_order(float("inf"))
 
     def test_inputs_less_than_one(self):
         with self.assertRaises(ValueError):
